move: feature 및 entities 폴더에 존재하는 auth 폴더 위치 변경 (#124)
