+++ apps/web/src/entities/auth/hooks/useAuth.ts
